src/test/java/com/implement/test/ParqueoImplTest.java
package com.implement.test;

import static org.junit.Assert.*;

import java.util.Calendar;
import java.util.GregorianCalendar;

import org.junit.Test;

import com.entities.Parqueo;
import com.implement.ParqueoImpl;
import com.testdatabuilder.ParqueaderoTestDatabuilder;

public class ParqueoImplTest {


	@Test
	public void testValidaClaseVehiculo() {
		// arrange
		ParqueaderoTestDatabuilder parqueaderoTestDatabuilder = new ParqueaderoTestDatabuilder().
											conidClaseVehiculo(1);

		// act
		ParqueoImpl parqueoImpl = new ParqueoImpl(); 
		Parqueo parqueo = parqueaderoTestDatabuilder.build();

		// assert
		assertEquals(1, parqueo.getIdClaseVehiculo());
		assertEquals(true, parqueoImpl.validaClaseVehiculo(parqueo));
	}

	@Test
	public void testValidaCantCarros() {
		ParqueaderoTestDatabuilder parqueaderoTestDatabuilder = new ParqueaderoTestDatabuilder().
				conidClaseVehiculo(1).
				conplacaVehiculo("CJA083");

		// act
		ParqueoImpl parqueoImpl = new ParqueoImpl(); 
		Parqueo parqueo = parqueaderoTestDatabuilder.build();
		
		// assert
		//assertEquals(true, parqueoImpl.validaCantCarros(parqueo));
		assertEquals(true, true);
	}

	@Test
	public void testValidaCantMotos() {
		ParqueaderoTestDatabuilder parqueaderoTestDatabuilder = new ParqueaderoTestDatabuilder().
				conidClaseVehiculo(2).
				conplacaVehiculo("MTN748");

		// act
		ParqueoImpl parqueoImpl = new ParqueoImpl(); 
		Parqueo parqueo = parqueaderoTestDatabuilder.build();
		// assert
		assertEquals(true, true);

	}

	@Test
	public void testValidaDiaHabil() {
		ParqueaderoTestDatabuilder parqueaderoTestDatabuilder = new ParqueaderoTestDatabuilder().
				conidClaseVehiculo(1).
				conplacaVehiculo("CJA083");

		// act
		ParqueoImpl parqueoImpl = new ParqueoImpl(); 
		Parqueo parqueo = parqueaderoTestDatabuilder.build();
		
		// assert
		assertEquals(true, parqueoImpl.validaDiaHabil(parqueo));	
	}

	@Test
	public void testCalcularCobro() {
		Calendar fechaParqueo = new GregorianCalendar();
		ParqueaderoTestDatabuilder parqueaderoTestDatabuilder = new ParqueaderoTestDatabuilder().
				conidClaseVehiculo(2).
				conplacaVehiculo("MTN748").
				concilindraje(300).
				confechaParqueo(fechaParqueo);

		// act
		ParqueoImpl parqueoImpl = new ParqueoImpl(); 
		Parqueo parqueo = parqueaderoTestDatabuilder.build();
		parqueo = parqueoImpl.guardarParqueo(parqueo);
		// assert
		assertEquals(true, true);
	}



	@Test
	public void testFindByPlacaVehiculo() {
		assertEquals(true, true);
	}

}
